Dockerfile: Use debian/stretch as base image
Ubuntu is based on debian and debian a more stable base to build on
IMHO.  This shouldn't effect much other than some minor changes in the
exact version clang we use to cross compile and the specific linked
host shared libraries.

Also, run `apt-get clean` to minimize image size (not that that really
matters since this is only used during SDK build).