Build and test SDK without requiring root (sudo) access (#143)
We still build with the `/opt/wasi-sdk` prefix but we don't actually
install into that directory.  Instead we install into `build/install`.

This means you don't need to run `sudo make` and you don't end up with
a all the build output being owned by root.

The main different is that this requires the `--sysroot` flag when
running clang to build the testsuite.