Add the `wasm32-wasi-pthread` output to the sysroot
Now that the `wasm32-wasi-pthread` target in wasi-libc has been worked
on for some amount of time, this change allows it to be released as a
part of wasi-sdk packages. It should be noted somewhere (a compiler
warning? README?) that this new target is not entirely stable, but that
can be added as a follow-on commit. This change is mainly focused on
ensuring that the build artifacts contain the new objects.