llvm: update to LLVM 16.0.0 (#314)
* llvm-project: update to 16.0.0 release

This changes updates the `src/llvm-project` submodule to the `HEAD` of
`release/16.x`, the same commit used to [release] the LLVM 16.0.0
binaries.

[release]: https://github.com/llvm/llvm-project/releases/tag/llvmorg-16.0.0

* fix: use only Clang's major version in install prefix

Due to [a change] in LLVM, Clang will expect to find the
`libclang_rt.builtins-wasm32.a` file in a path that only contains the
major version (`16`) instead of the entire version (`16.0.0`) as was
previously the case. This change modifies the `CMAKE_INSTALL_PREFIX` to
use Clang's major version only.

[a change]: https://reviews.llvm.org/D125860

* review: only use `llvm_version_major.sh`

Since the `Makefile` can get by with only knowing Clang's major version,
this change removes `llvm_version.sh` and sets `CLANG_VERSION` to use
only the major part.