move `CMAKE_POSITION_INDEPENDENT_CODE` to Makefile
This is where I originally meant the directive to go, but it didn't seem to work
the first time I tried it, hence the added lines in wasi-sdk.cmake and
wasi-sdk-pthread.cmake.  However, it seems to work fine now, so I'm switching it
back to be consistent with other similar CMake flags.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>