build and test `wasm32-wasi-preview2` target
This updates `wasi-libc` to include
https://github.com/WebAssembly/wasi-libc/pull/457, which adds preliminary
support for the new `wasm32-wasi-preview2` target.

It also adds support for testing the new target using Wasmtime 16.0.0 and
`wit-component`.  Note that Wasmtime produces different output when reporting
errors for Preview 2 components than it does for Preview 1 modules, so I've
added a few .expected files specific to Preview 2.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>