Update LLVM to 18.1.2 and add wasm-component-ld (#402)
* Update LLVM to 18.1.2 and add wasm-component-ld

This commit has two intertwined changes within it. First the LLVM
submodule is updated to the 18.1.2 release branch. This alone would
cause CI and tests to fail due to differing behavior for the
`wasm32-wasip2` target. To fix these test failures the
`wasm-component-ld` tool is added to the build. This tool is a
Rust-written tool and installed via `cargo install` as part of the build
at a pinned version written in the `Makefile`. This linker, used for
components, is then used for the `wasm32-wasip2` target.

Tests and CI are then updated to skip the need to have `wasm-tools` or
the adapter for WASI when making components since that's now the job of
`wasm-component-ld`. This then necessitated some changes to tests too.

* Add back accidentally deleted CI

* Inherit tools on PATH on Windows

* Add Rust to docker build

* Put rust in a different location in docker

* Try to fix permissions

* Review comments

* Revert changes to test outputs