Refactor handling env vars in CI build (#455)
* Refactor handling env vars in CI build

This commit updates the handling of various environment variables to
centralize them in one location in `main.yml` as part of CI
configuration. This removes a few custom-named variables in favor of
using explicitly-listed env vars. Additionally this moves some env var
management from the CI container files to the github actions config as
well to keep everything in one place ideally.

* Add some more comments about where configuration comes in