Publish header with toolchain versions
This change is an attempt to resolve [wasi-libc#490] but at the wasi-sdk
level. It adds a `version.h` header file to
`share/wasi-sysroot/include/<target>/wasi` so that users have
programmatic access to some extra information to output better error
messages, e.g. This `version.h` looks something like:

  ```c
  // Generated by wasi-sdk's `version.py` script.
  #ifndef VERSION_H
  #define VERSION_H

  #define WASI_SDK_VERSION "24.6g754aec3d6f58+m"
  #define WASI_LIBC_VERSION "b9ef79d7dbd4"

  #endif
  ```

It _is_ a bit strange that we're adding to wasi-libc's include files
from wasi-sdk (it would be cleaner if it happened in wasi-libc directly)
but wasi-sdk actually has the information available.

[wasi-libc#490]: https://github.com/WebAssembly/wasi-libc/issues/490