Remove `wasm32-wasi` and `wasm32-wasi-threads` targets
These are replaced by `wasm32-wasip1` and `wasm32-wasip1-threads`,
respectively. The intention is to free up the namespace of "wasm32-wasi"
for a hypothetical future target of WASI 1.0. This naming matches the
Rust naming scheme which dropped support for `wasm32-wasi` some time
ago (and renamed it to `wasm32-wasip1`)